removed code related to units. FOr some reason it segfaults with Python3. This is a c code and windowsglob function written in old-style c also were segfaulting. Will look closer at this issue later. This is not essential code anyway

+++ CompuCell3D/core/BasicUtils/BasicArray.h
@@ -85,7 +85,7 @@ class BasicArray :public std::vector<T>{
   /** 
    * @return The current size of the array.
    */
-  unsigned int getSize() const {return std::vector<T>::size();}
+  size_t getSize() const {return std::vector<T>::size();}
 
   /** 
    * setSize() will increase the array capacity if
